INS-979: Move lifeline index to lifeline file.
